Changed TKL price behavior and URL
