+++ src/components/Dashboard/Portfolio/Portfolio.tsx
@@ -1,8 +1,7 @@
-import React, { ReactElement, useEffect } from 'react';
+import React, { ReactElement } from 'react';
 import { useSelector } from 'react-redux';
 
 import styled from '@emotion/styled';
-import axios from 'axios';
 
 import { dispatch } from 'store/rematch';
 import {
@@ -13,7 +12,6 @@ import {
   selectUnspentBalance,
 } from 'store/selectors';
 import { V } from 'util/theming';
-import { TOKEL_PRICE_UPDATE_PERIOD_MS, TOKEL_PRICE_URL } from 'vars/defines';
 
 import { WidgetContainer } from '../widgets/common';
 import PortfolioItem from './PortfolioItem';
@@ -30,15 +28,6 @@ const PortfolioRoot = styled(WidgetContainer)`
   border-radius: ${V.size.borderRadius};
 `;
 
-const fetchTokelPrice = async () => {
-  try {
-    const priceJson = await axios(TOKEL_PRICE_URL);
-    dispatch.environment.SET_TOKEL_PRICE_USD(priceJson.data[0]?.price);
-  } catch (e) {
-    console.log(e);
-  }
-};
-
 const Portfolio = (): ReactElement => {
   const currentAsset = useSelector(selectCurrentAsset);
   const balance = useSelector(selectUnspentBalance);
@@ -47,14 +36,6 @@ const Portfolio = (): ReactElement => {
 
   const tokelPriceUSD = useSelector(selectTokelPriceUSD);
   const priceString = tokelPriceUSD ? ` ~ $${Math.round(balance * tokelPriceUSD * 100) / 100}` : '';
-
-  useEffect(() => {
-    fetchTokelPrice();
-    const priceClock = setInterval(fetchTokelPrice, TOKEL_PRICE_UPDATE_PERIOD_MS);
-    return () => {
-      clearInterval(priceClock);
-    };
-  }, []);
 
   return (
     <PortfolioRoot>
