+++ src/vars/defines.ts
@@ -13,8 +13,8 @@ export enum IpfsAction {
 // TODO move to user settings?
 export const DEFAULT_IPFS_FALLBACK_GATEWAY = 'https://ipfs.io/ipfs';
 
-export const TOKEL_PRICE_URL = `https://api.nomics.com/v1/currencies/ticker?key=${process.env.NOMICS_API_KEY}&ids=TKL2`;
-export const TOKEL_PRICE_UPDATE_PERIOD_MS = 10_000;
+export const TOKEL_PRICE_URL = 'http://price.tokel.io';
+export const TOKEL_PRICE_UPDATE_PERIOD_MS = 7_200_000; // two hours
 
 export const TOPBAR_HEIGHT_PX = 38;
 export const PORTFOLIO_ITEM_HEIGHT_PX = 86;
